esplink: fix overflow on load percentage for large Linux images

+++ utils/esplink/src/edcl.c
@@ -1,14 +1,14 @@
 #include "edcl.h"
 
 // Helper functions
-static void print_progress(u32 progress, u32 total, const char *prefix)
+static void print_progress(u64 progress, u64 total, const char *prefix)
 {
 	const u32 symbols = 40;
 	int i;
 
-	u32 percent = progress * 100 / total;
+	u64 percent = progress * 100 / total;
 
-	u32 fraction = progress * symbols / total;
+	u64 fraction = progress * symbols / total;
 
 	printf("%s: [", prefix);
 
@@ -17,7 +17,7 @@ static void print_progress(u32 progress, u32 total, const char *prefix)
 		printf("%c", c);
 	}
 
-	printf("] %d%%", percent);
+	printf("] %lld%%", percent);
 	if (progress == total)
 		printf("\n");
 	else
